Extract token class and create package for tokenization
